refactoring add set into small child component
